requirements.txt
django==1.5.1
BeautifulSoup4==3.2.1
html5lib==1.0b3
django-filebrowser==3.5.2

